Fix #58 refactor api structs (#72)
