+++ pkg/cybr/api/responses/listaccounts.go
@@ -1,0 +1,7 @@
+package responses
+
+// ListAccount response from listing accounts
+type ListAccount struct {
+	Value []GetAccount `json:"value"`
+	Count int          `json:"count"`
+}
