+++ pkg/cybr/api/users.go
@@ -5,50 +5,17 @@ import (
 	"fmt"
 	"net/url"
 
+	"github.com/infamousjoeg/cybr-cli/pkg/cybr/api/queries"
+	"github.com/infamousjoeg/cybr-cli/pkg/cybr/api/requests"
+	"github.com/infamousjoeg/cybr-cli/pkg/cybr/api/responses"
 	httpJson "github.com/infamousjoeg/cybr-cli/pkg/cybr/helpers/httpjson"
 )
-
-// UnsuspendUserRequest request used when unsuspending user
-type UnsuspendUserRequest struct {
-	Suspended bool `json:"Suspended"`
-}
-
-// ListUsersQueryParams represents valid query parameters when listing users
-type ListUsersQueryParams struct {
-	Search string `query_key:"search"`
-	Filter string `query_key:"filter"`
-}
-
-// ListUsersResponse response when listing users
-type ListUsersResponse struct {
-	Users []UserResponse `json:"Users"`
-	Total int            `json:"Total"`
-}
-
-// UserResponse represents one user in ListUsersResponse
-type UserResponse struct {
-	ID                 int                     `json:"id"`
-	Username           string                  `json:"username"`
-	Source             string                  `json:"source"`
-	UserType           string                  `json:"userType"`
-	ComponentUser      bool                    `json:"componentUser"`
-	VaultAuthorization []string                `json:"vaultAuthorization"`
-	Location           string                  `json:"location"`
-	PersonalDetails    PersonalDetailsResponse `json:"personalDetails"`
-}
-
-// PersonalDetailsResponse represents one users personal details
-type PersonalDetailsResponse struct {
-	FirstName  string `json:"firstName"`
-	MiddleName string `json:"middleName"`
-	LastName   string `json:"lastName"`
-}
 
 // UnsuspendUser activates a suspended user. It does not activate an inactive user.
 func (c Client) UnsuspendUser(username string) error {
 	url := fmt.Sprintf("%s/PasswordVault/WebServices/PIMServices.svc/Users/%s", c.BaseURL, url.QueryEscape(username))
 
-	body := UnsuspendUserRequest{
+	body := requests.UnsuspendUser{
 		Suspended: false,
 	}
 
@@ -61,17 +28,17 @@ func (c Client) UnsuspendUser(username string) error {
 }
 
 // ListUsers returns a list of all existing users in the Vault except for the Master and the Batch built-in users.
-func (c Client) ListUsers(query *ListUsersQueryParams) (ListUsersResponse, error) {
+func (c Client) ListUsers(query *queries.ListUsers) (responses.ListUsers, error) {
 	url := fmt.Sprintf("%s/PasswordVault/api/Users%s", c.BaseURL, httpJson.GetURLQuery(query))
 
 	response, err := httpJson.Get(url, c.SessionToken, c.InsecureTLS, c.Logger)
 	if err != nil {
 		returnedError, _ := json.Marshal(response)
-		return ListUsersResponse{}, fmt.Errorf("Failed to list users. %s. %s", string(returnedError), err)
+		return responses.ListUsers{}, fmt.Errorf("Failed to list users. %s. %s", string(returnedError), err)
 	}
 
 	jsonString, _ := json.Marshal(response)
-	ListUsersResponse := ListUsersResponse{}
+	ListUsersResponse := responses.ListUsers{}
 	err = json.Unmarshal(jsonString, &ListUsersResponse)
 	return ListUsersResponse, err
 }
